changes in UI

+++ src/main/java/com/i2i/vehicleloan/dao/LoanDetailDao.java
@@ -57,5 +57,5 @@ public interface LoanDetailDao extends GenericDao<LoanDetail, Long> {
      * @throws DatabaseException
      *     It handle all the custom exception in vehicle loan application.
      */
-    public List<LoanDetail> retrieveLoanDetailByUserId(int userId) throws DatabaseException;
+    public List<LoanDetail> retrieveLoanDetailByUserId(Long userId) throws DatabaseException;
 }
