+++ src/main/java/com/i2i/vehicleloan/service/impl/LoanDetailServiceImpl.java
@@ -81,7 +81,7 @@ public class LoanDetailServiceImpl extends GenericManagerImpl<Company, Long> imp
      * @throws DatabaseException
      *     It handle all the custom exception in vehicle loan application. 
      */
-    public LoanDetail retrieveLoanDetailByUserId(int userId)throws DatabaseException {
+    public LoanDetail retrieveLoanDetailByUserId(Long userId)throws DatabaseException {
         return retrieveLatestLoanDetailFromList(loanDetailDao.retrieveLoanDetailByUserId(userId));
     }
